requirements.txt
streamlit
yfinance
numpy
pandas
python-dotenv
google-generativeai
